Extraction module
The Extraction module offers an API to extract the
features from a given music.
The module uses the Essentia extractor
`streaming_extractor_music` with a given profile to extract
the features from the music.
After the extraction, a parse is made to ensure the use
of only the relevant features to the player.

Resolves #13